cvsps: should ignore TRUNK branch if it exists in log
On anoncvs.opensource.apple.com (where access to the now-open-source
WebKit is provided), some files have a branch named "TRUNK" with a
revision of "1".  Yes, I know that's not supposed to be possible, but
they guys at Apple found a way to do it.  Apparently, using "TRUNK"
works in places that "HEAD" does not for some cvs commands.

The current version (2.1) will die when it hits the "TRUNK: 1"
branch/revision.  The attached patch will ignore any branch or tag named
"TRUNK".