Report correct ancestor branch for file added on branch
I tried the git-cvsimport script and found that in the
resulting git tree, on a branch-off-a-branch, files were missing from
the branch if they had been added to the branch's immediate parent.  I
really wanted this to work better, so I looked into it and learned
that CVSps was incorrectly reporting that the "ancestor" branch was
HEAD for a branch-off-a-branch.

Since CVSps wasn't reporting the correct branch relationships, I
couldn't blame git-cvsimport.  Trying to improve CVSps to get better
parent-detection proved much harder than I expected.  I had to use a
completely different algorithm that loops over all files instead of
just the files in the patchsets that are on the child branch.

Now that I've done that, [David, please see attached patch] CVSps does
report more accurate branch parents.  But, testing git-cvsimport with
the corrected CVSps output shows that it *still* doesn't produce the
right tree.  In fact, it's WORSE!  In the simple test script, not only
is file 'b' not on the 'dev' branch, but the 'stable' branch and the
file 'b' is completely MISSING form the git tree!