Semicolons are now escaped inside file names stored in the cache.
Previously, if a file name contained a semicolon, cvsps choked when
reading information from cache. The problem was noticed when importing
Eclipse DLTK project repository.

Now semicolons are escaped by prepending a backslash to them.
(Backslashes are not escaped, this would involve more work and
I hope backslashes are much less likely to be used as part of
a file name.)